Added mode function
Copied over the jwsinfo and jwssetter files from JWS which I've just recently
added the feature to. Updated the config window to read it from the file, set
the file, and also so that it correctly sets the wallpaper with the current
window's mode option when using the try function.